Malipo/sms_sender_project/requirements.txt
africastalking==1.2.6
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
